+++ src/Console/Application.php
@@ -19,7 +19,6 @@ use CacheTool\CacheTool;
 use CacheTool\Command as CacheToolCommand;
 use CacheTool\Monolog\ConsoleHandler;
 use Monolog\Logger;
-use SelfUpdate\SelfUpdateCommand;
 use Symfony\Component\Console\Application as BaseApplication;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputOption;
@@ -61,7 +60,11 @@ class Application extends BaseApplication
     protected function getDefaultCommands()
     {
         $commands = parent::getDefaultCommands();
-        $commands[] = new SelfUpdateCommand('gordalina/cachetool', '@package_version@', 'gordalina/cachetool');
+        $commands[] = new CacheToolCommand\SelfUpdateCommand(
+            'gordalina/cachetool',
+            '@package_version@',
+            'gordalina/cachetool'
+        );
 
         if (in_array('apcu', $this->config['extensions'], true)) {
             $commands[] = new CacheToolCommand\ApcuCacheClearCommand();